Fix loading tests in Makefile
Only explicitly run the tests in the TEST variable in the Makefile.
Also, make all paths in the `cask exec` command absolute.

This works around an issue where a `flycheck_org-gcal-test.el` file
somehow committed to the repo was overriding `org-gcal-test.el`.

To further avoid this issue, remove that file and ignore it in the
future.